// src/models/enums/GameStatCodeEnum.ts
export enum GameStatCodeEnum {
    G = "G", // Goal
    M = "M", // Miss
    PM = "PM", //Penalty Miss
    S = "S", // Subtract
    P = "P", // Points
    MP = "MP", // Missed Points
    F = "F", // Foul
    TC = "TC", // Timer Change
    AT = "AT", // Additional Time
    Pe = "Pe", // Penalty
    OG = "OG", // Own Goal
    YC = "YC", // Yellow Card
    RC = "RC", // Red Card
    A = "A", // Assist
}